Move test cases from RecoveryTests.swift to Declaration.swift
These tests now also use `AssertParse` and should thus live next to all the other declaration parsing tests.
